+++ src/us/kbase/workspace/test/listener/BadListenerFactory.java
@@ -62,6 +62,19 @@ public class BadListenerFactory implements WorkspaceEventListenerFactory {
 
 		@Override
 		public void revertObject(long workspaceId, long objectId, int version) {}
+
+		@Override
+		public void setObjectDeleted(long workspaceId, long objectId, boolean delete) {}
+
+		@Override
+		public void copyObject(
+				long workspaceId,
+				long objectId,
+				int version,
+				boolean allVersionsCopied) {}
+
+		@Override
+		public void saveObject(long workspaceId, long objectId, int version, String type) {}
 		
 	}